requirements.txt
torch==2.11.0
torchvision==0.26.0
efficientnet-pytorch==0.7.1
timm==0.9.12
numpy>=1.26.4
pandas>=2.2.1
Pillow>=10.2.0
scikit-learn>=1.4.1
matplotlib>=3.8.3
seaborn>=0.13.2
opencv-python>=4.9.0.80
mlflow>=2.11.1
hydra-core>=1.3.2
omegaconf>=2.3.0
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.4
onnx>=1.16.0
onnxruntime>=1.17.1
onnxscript>=0.1.0
kaggle>=1.6.6
tqdm>=4.66.2
httpx>=0.27.0
loralib>=0.1.1